pyproject.toml
[tool.poetry]
name = "playwright-python-samples"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "playwright_python_samples"}]

[tool.poetry.dependencies]
python = "^3.11"
pytest-playwright = "^0.3.0"
pytest = "^7.2.0"


[tool.poetry.group.dev.dependencies]
black = "^22.12.0"
isort = "^5.10.1"
mypy = "^0.991"
pylint = "^2.15.8"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pylint.'MESSAGES CONTROL']
disable = [
    "missing-function-docstring",
    "missing-class-docstring",
    "missing-module-docstring"
]

[tool.isort]
profile = "black"